+++ examples/modal-data-router/vite.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from "vite";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  server: {
-    port: 3000,
-  },
-});