+++ packages/react-router/__tests__/dom/ssr/meta-test.tsx
@@ -279,6 +279,48 @@ describe("meta", () => {
     );
   });
 
+  it("{ 'script:ld+json': [...] } adds a <script type='application/ld+json' /> with an array of objects", () => {
+    let jsonLd = [
+      {
+        "@context": "http://schema.org",
+        "@type": "Organization",
+        name: "Acme Inc.",
+        url: "https://acme.example.com",
+      },
+      {
+        "@context": "http://schema.org",
+        "@type": "BreadcrumbList",
+        itemListElement: [
+          {
+            "@type": "ListItem",
+            position: 1,
+            name: "Home",
+            item: "https://acme.example.com",
+          },
+        ],
+      },
+    ];
+
+    let RoutesStub = createRoutesStub([
+      {
+        path: "/",
+        meta: () => [
+          {
+            "script:ld+json": jsonLd,
+          },
+        ],
+        Component: Meta,
+      },
+    ]);
+
+    let { container } = render(<RoutesStub />);
+
+    let scriptTagContents =
+      container.querySelector('script[type="application/ld+json"]')
+        ?.innerHTML || "{}";
+    expect(JSON.parse(scriptTagContents)).toEqual(jsonLd);
+  });
+
   it("{ tagName: 'link' } adds a <link />", () => {
     let RoutesStub = createRoutesStub([
       {