+++ packages/react-router/lib/components.tsx
@@ -398,7 +398,7 @@ export interface RouterProviderProps {
    * and is only present for render errors.
    *
    * ```tsx
-   * <RouterProvider onError=(error, info) => {
+   * <RouterProvider onError={(error, info) => {
    *   let { location, params, pattern, errorInfo } = info;
    *   console.error(error, location, errorInfo);
    *   reportToErrorService(error, location, errorInfo);