+++ packages/create-react-router/CHANGELOG.md
@@ -1,5 +1,17 @@
 # `create-react-router`
 
+## v7.17.0
+
+### Patch Changes
+
+- _No changes_
+
+## v7.16.0
+
+### Patch Changes
+
+- _No changes_
+
 ## v7.15.1
 
 ### Patch Changes