+++ packages/react-router-fs-routes/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `@react-router/fs-routes`
 
+## v7.17.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - [`@react-router/dev@7.17.0`](https://github.com/remix-run/react-router/releases/tag/@react-router/dev@7.17.0)
+
+## v7.16.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - [`@react-router/dev@7.16.0`](https://github.com/remix-run/react-router/releases/tag/@react-router/dev@7.16.0)
+
 ## v7.15.1
 
 ### Patch Changes